Remove reducers folder
